pyproject.toml
[project]
name = "drf-catalyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework>=3.15.2",
    "django-environ>=0.11.2",
    "django-phonenumber-field>=8.0.0",
    "djangorestframework-simplejwt>=5.4.0",
    "drf-yasg>=1.21.8",
    "django-filter>=24.3",
    "phonenumbers>=8.13.52",
    "pillow>=11.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
]
